Fjernet modalvisning av feilmeldinger for sakstype eu/eøs og folketrygdloven. Disse vises nå som en rekke Nav.Alertstripe under Stegvelger
