+++ src/ducks/feiletrespons/actions.ts
@@ -1,0 +1,5 @@
+import * as Types from "./types";
+
+export function resetFeiletresponsState() {
+  return { type: Types.RESET };
+}
