+++ src/ducks/feiletrespons/index.ts
@@ -1,7 +1,8 @@
 import * as feiletresponsSelectors from "./selectors";
+import * as feiletresponsOperations from "./operations";
 
 import feiletresponsReducers from "./reducers";
 
-export { feiletresponsSelectors };
+export { feiletresponsSelectors, feiletresponsOperations };
 
 export default feiletresponsReducers;
