+++ src/ducks/feiletrespons/operations.ts
@@ -1,0 +1,5 @@
+import * as Actions from "./actions";
+
+export function resetFeiletresponsState() {
+  return Actions.resetFeiletresponsState();
+}
