+++ src/ducks/modaler/types.js
@@ -2,6 +2,5 @@ export const OPPDATER_HENLEGG = "modaler/OPPDATER_HENLEGG";
 export const OPPDATER_AVSLUTT_SAK_SOM_BORTFALT = "modaler/OPPDATER_AVSLUTT_SAK_SOM_BORTFALT";
 export const OPPDATER_AVSLAG_SOKNAD = "modaler/OPPDATER_AVSLAG_SOKNAD";
 export const OPPDATER_OPPFRISK = "modaler/OPPDATER_OPPFRISK";
-export const OPPDATER_VALIDERING = "modaler/OPPDATER_VALIDERING";
 export const OPPDATER_REVURDER_FAGSAK = "modaler/OPPDATER_REVURDER_FAGSAK";
 export const OPPDATER_FERDIGBEHANDLE_NY_VURDERING = "modaler/OPPDATER_FERDIGBEHANDLE_NY_VURDERING";
