+++ src/felleskomponenter/dialogboks/index.ts
@@ -4,7 +4,7 @@ import DialogboksFerdigbehandleNyVurdering from "./ferdigbehandleNyVurdering";
 import DialogboksHenleggSak from "./henlegg";
 import DialogboksOppfriskSak from "./oppfrisk";
 import DialogboksRevurderFagsak from "./revurderFagsak";
-import DialogboksValidering from "./validering";
+import FeilmeldingDialog from "./validering";
 
 export {
   DialogboksAvslagSoknad,
@@ -13,5 +13,5 @@ export {
   DialogboksHenleggSak,
   DialogboksOppfriskSak,
   DialogboksRevurderFagsak,
-  DialogboksValidering,
+  FeilmeldingDialog,
 };
