+++ src/felleskomponenter/dialogboks/validering/feilmeldingDialog.less
@@ -1,6 +1,6 @@
 @import "../../../constants";
 
-.dialogboksValidering {
+.feilmeldingDialog {
   width: 50%;
   display: flex;
   justify-content: center;
