+++ src/felleskomponenter/stegvelger/stegvelger.less
@@ -45,4 +45,8 @@
       background-image: url("../../resources/images/ikon-steg-feil.svg");
     }
   }
+
+  .varsel {
+    margin-bottom: 0.5em;
+  }
 }
