+++ src/sider/journalforing/journalforing.js
@@ -17,7 +17,7 @@ import Sticky from "../../felleskomponenter/sticky";
 import PDFDokument from "./komponenter/pdfdokument";
 import JournalforingSED from "./komponenter/journalforingsed";
 import JournalforingForm from "./komponenter/journalforingform";
-import { DialogboksValidering } from "../../felleskomponenter/dialogboks";
+import { FeilmeldingDialog } from "../../felleskomponenter/dialogboks";
 
 import { journalforingOperations, journalforingSelectors } from "../../ducks/journalforing";
 import { formSelectors } from "../../ducks/form";
@@ -31,9 +31,9 @@ import "./journalforing.css";
 class Journalforing extends Component {
   state = {
     valgtDokumentID: -1,
-    visDialogboksValidering: false,
-    dialogboksValideringFeilmeldinger: [],
-    dialogboksValideringValideringer: [],
+    visFeilmeldingDialog: false,
+    feilmeldinger: [],
+    valideringer: [],
   };
   async componentDidMount() {
     const { journalpostID } = this.props.match.params;
@@ -143,26 +143,26 @@ class Journalforing extends Component {
 
   resetFeilmeldinger = () => {
     this.setState({
-      dialogboksValideringFeilmeldinger: [],
-      dialogboksValideringValideringer: [],
+      feilmeldinger: [],
+      valideringer: [],
     });
   };
 
-  skjulDialogboksValideringOgResetFeilmeldinger = () => {
-    this.setState({ visDialogboksValidering: false });
+  skjulFeilmeldingDialogOgResetFeilmeldinger = () => {
+    this.setState({ visFeilmeldingDialog: false });
     this.resetFeilmeldinger();
   };
 
   sjekkErrorOgVisFeilmelding = (error) => {
     if (error.body.feilkoder?.length > 0) {
       this.setState({
-        dialogboksValideringValideringer: error.body.feilkoder,
+        valideringer: error.body.feilkoder,
       });
     }
 
     if (error.status >= 500) {
       this.setState({
-        dialogboksValideringFeilmeldinger: [
+        feilmeldinger: [
           {
             tittel: "Teknisk feil",
             innhold:
@@ -172,7 +172,7 @@ class Journalforing extends Component {
       });
     } else if (error.status >= 400) {
       this.setState({
-        dialogboksValideringFeilmeldinger: [
+        feilmeldinger: [
           {
             tittel: "Feil",
             innhold: error.body.message,
@@ -181,7 +181,7 @@ class Journalforing extends Component {
       });
     }
 
-    this.setState({ visDialogboksValidering: true });
+    this.setState({ visFeilmeldingDialog: true });
   };
 
   /** Når saksbehandler klikker "knytt til eksisterende sak" skal det åpnes for validering av
@@ -223,7 +223,7 @@ class Journalforing extends Component {
 
     try {
       await Api.Journalforing.tilordne(journalforingData);
-      this.setState({ visDialogboksValidering: false });
+      this.setState({ visFeilmeldingDialog: false });
       return tilForsiden();
     } catch (error) {
       this.sjekkErrorOgVisFeilmelding(error);
@@ -283,7 +283,7 @@ class Journalforing extends Component {
 
     try {
       await Api.Journalforing.opprett(journalforingData);
-      this.setState({ visDialogboksValidering: false });
+      this.setState({ visFeilmeldingDialog: false });
       return tilForsiden();
     } catch (error) {
       this.sjekkErrorOgVisFeilmelding(error);
@@ -414,7 +414,7 @@ class Journalforing extends Component {
 
     try {
       await Api.Journalforing.sed(data);
-      this.setState({ visDialogboksValidering: false });
+      this.setState({ visFeilmeldingDialog: false });
       return tilForsiden();
     } catch (error) {
       this.sjekkErrorOgVisFeilmelding(error);
@@ -450,7 +450,7 @@ class Journalforing extends Component {
       settFeltInnhold,
     } = this.props;
 
-    const { visDialogboksValidering, dialogboksValideringFeilmeldinger, dialogboksValideringValideringer } = this.state;
+    const { visFeilmeldingDialog, feilmeldinger, valideringer } = this.state;
 
     const { knyttTilEksisterendeSak, opprettFagsak, hentOgVisAvsender, hentOgVisBruker, hentOgVisRepresentant } = this;
     const { journalpostID } = this.props.match.params;
@@ -537,11 +537,11 @@ class Journalforing extends Component {
             </Nav.Row>
           </Nav.Container>
         </div>
-        {visDialogboksValidering && (
-          <DialogboksValidering
-            avbryt={this.skjulDialogboksValideringOgResetFeilmeldinger}
-            feilmeldinger={dialogboksValideringFeilmeldinger}
-            valideringer={dialogboksValideringValideringer}
+        {visFeilmeldingDialog && (
+          <FeilmeldingDialog
+            avbryt={this.skjulFeilmeldingDialogOgResetFeilmeldinger}
+            feilmeldinger={feilmeldinger}
+            valideringer={valideringer}
           />
         )}
       </>
